pyproject.toml
[project]
name = "myactuator-custom"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.5",
    "numpy>=2.2.6",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "pynput>=1.8.1",
    "python-can>=4.5.0",
    "myactuator-rmd-py",
]

[tool.uv.sources]
myactuator-rmd-py = { path = "utils/myactuator_rmd" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
